+++ x17/x17.cu
@@ -78,7 +78,7 @@ extern void x17_haval256_cpu_hash_64(int thr_id, uint32_t threads, uint32_t star
 
 extern void quark_compactTest_cpu_init(int thr_id, uint32_t threads);
 extern void quark_compactTest_cpu_hash_64(int thr_id, uint32_t threads, uint32_t startNounce, uint32_t *inpHashes,
-											uint32_t *d_noncesTrue, size_t *nrmTrue, uint32_t *d_noncesFalse, size_t *nrmFalse,
+											uint32_t *d_noncesTrue, uint32_t *nrmTrue, uint32_t *d_noncesFalse, uint32_t *nrmFalse,
 											int order);
 
 // X17 Hashfunktion
